rewrite main models
