+++ api/admin.py
@@ -1,15 +1,15 @@
 from django.contrib import admin
-from django.contrib.gis.admin import OSMGeoAdmin
-from .models import *
+# from django.contrib.gis.admin import OSMGeoAdmin
+# from .models import *
 
-# Register your models here.
-
-
-class IssueAdmin(OSMGeoAdmin):
-    list_display = ("full_name", "active", "verified",
-                    "created", "updated", "creator")
-    list_filter = ("active", "verified", "updated")
+# # Register your models here.
 
 
-admin.site.register(Issue, IssueAdmin)
-admin.site.register(Statistics)
+# class IssueAdmin(OSMGeoAdmin):
+#     list_display = ("full_name", "active", "verified",
+#                     "created", "updated", "creator")
+#     list_filter = ("active", "verified", "updated")
+
+
+# admin.site.register(Issue, IssueAdmin)
+# admin.site.register(Statistics)
