+++ api/serializer.py
@@ -2,7 +2,7 @@ from django.db.models import fields
 from rest_framework import serializers
 from rest_framework.validators import UniqueValidator
 from django.contrib.auth.password_validation import validate_password
-from .models import Issue
+from server_models.models import Issue
 from django.contrib.auth.models import User
 import datetime
 from django.utils import timezone
