+++ server_models/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ModelsConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'server_models'
